+++ Lib/IntCodeTest/ExecutorTests.cs
@@ -15,17 +15,17 @@ namespace IntCodeTest
       long expected2 = 1;
       long expected3 = 0;
 
-      executor.InputQueue.Enqueue((long)7);
+      executor.InputQueue.Enqueue(7);
       executor.Execute();
-      var actual1 = (long)executor.OutputQueue.Dequeue();
+      var actual1 = executor.OutputQueue.Dequeue();
       executor.Initialize();
-      executor.InputQueue.Enqueue((long)8);
+      executor.InputQueue.Enqueue(8);
       executor.Execute();
-      var actual2 = (long)executor.OutputQueue.Dequeue();
+      var actual2 = executor.OutputQueue.Dequeue();
       executor.Initialize();
-      executor.InputQueue.Enqueue((long)15);
+      executor.InputQueue.Enqueue(15);
       executor.Execute();
-      var actual3 = (long)executor.OutputQueue.Dequeue();
+      var actual3 = executor.OutputQueue.Dequeue();
 
       Assert.AreEqual<long>(expected1, actual1, "OK!");
       Assert.AreEqual<long>(expected2, actual2, "OK!");
@@ -41,17 +41,17 @@ namespace IntCodeTest
       long expected2 = 1;
       long expected3 = 0;
 
-      executor.InputQueue.Enqueue((long)3);
+      executor.InputQueue.Enqueue(3);
       executor.Execute();
-      var actual1 = (long)executor.OutputQueue.Dequeue();
+      var actual1 = executor.OutputQueue.Dequeue();
       executor.Initialize();
-      executor.InputQueue.Enqueue((long)8);
+      executor.InputQueue.Enqueue(8);
       executor.Execute();
-      var actual2 = (long)executor.OutputQueue.Dequeue();
+      var actual2 = executor.OutputQueue.Dequeue();
       executor.Initialize();
-      executor.InputQueue.Enqueue((long)9);
+      executor.InputQueue.Enqueue(9);
       executor.Execute();
-      var actual3 = (long)executor.OutputQueue.Dequeue();
+      var actual3 = executor.OutputQueue.Dequeue();
 
       Assert.AreEqual<long>(expected1, actual1, "OK!");
       Assert.AreEqual<long>(expected2, actual2, "OK!");
@@ -67,17 +67,17 @@ namespace IntCodeTest
       long expected2 = 0;
       long expected3 = 0;
 
-      executor.InputQueue.Enqueue((long)7);
+      executor.InputQueue.Enqueue(7);
       executor.Execute();
-      var actual1 = (long)executor.OutputQueue.Dequeue();
+      var actual1 = executor.OutputQueue.Dequeue();
       executor.Initialize();
-      executor.InputQueue.Enqueue((long)8);
+      executor.InputQueue.Enqueue(8);
       executor.Execute();
-      var actual2 = (long)executor.OutputQueue.Dequeue();
+      var actual2 = executor.OutputQueue.Dequeue();
       executor.Initialize();
-      executor.InputQueue.Enqueue((long)11);
+      executor.InputQueue.Enqueue(11);
       executor.Execute();
-      var actual3 = (long)executor.OutputQueue.Dequeue();
+      var actual3 = executor.OutputQueue.Dequeue();
 
       Assert.AreEqual<long>(expected1, actual1, "OK!");
       Assert.AreEqual<long>(expected2, actual2, "OK!");
@@ -93,17 +93,17 @@ namespace IntCodeTest
       long expected2 = 0;
       long expected3 = 0;
 
-      executor.InputQueue.Enqueue((long)7);
+      executor.InputQueue.Enqueue(7);
       executor.Execute();
-      var actual1 = (long)executor.OutputQueue.Dequeue();
+      var actual1 = executor.OutputQueue.Dequeue();
       executor.Initialize();
-      executor.InputQueue.Enqueue((long)8);
+      executor.InputQueue.Enqueue(8);
       executor.Execute();
-      var actual2 = (long)executor.OutputQueue.Dequeue();
+      var actual2 = executor.OutputQueue.Dequeue();
       executor.Initialize();
-      executor.InputQueue.Enqueue((long)11);
+      executor.InputQueue.Enqueue(11);
       executor.Execute();
-      var actual3 = (long)executor.OutputQueue.Dequeue();
+      var actual3 = executor.OutputQueue.Dequeue();
 
       Assert.AreEqual<long>(expected1, actual1, "OK!");
       Assert.AreEqual<long>(expected2, actual2, "OK!");
@@ -119,17 +119,17 @@ namespace IntCodeTest
       long expected2 = 1;
       long expected3 = 1;
 
-      executor.InputQueue.Enqueue((long)0);
+      executor.InputQueue.Enqueue(0);
       executor.Execute();
-      var actual1 = (long)executor.OutputQueue.Dequeue();
+      var actual1 = executor.OutputQueue.Dequeue();
       executor.Initialize();
-      executor.InputQueue.Enqueue((long)1);
+      executor.InputQueue.Enqueue(1);
       executor.Execute();
-      var actual2 = (long)executor.OutputQueue.Dequeue();
+      var actual2 = executor.OutputQueue.Dequeue();
       executor.Initialize();
-      executor.InputQueue.Enqueue((long)9);
+      executor.InputQueue.Enqueue(9);
       executor.Execute();
-      var actual3 = (long)executor.OutputQueue.Dequeue();
+      var actual3 = executor.OutputQueue.Dequeue();
 
       Assert.AreEqual<long>(expected1, actual1, "OK!");
       Assert.AreEqual<long>(expected2, actual2, "OK!");
@@ -145,17 +145,17 @@ namespace IntCodeTest
       long expected2 = 0;
       long expected3 = 1;
 
-      executor.InputQueue.Enqueue((long)-1);
+      executor.InputQueue.Enqueue(-1);
       executor.Execute();
-      var actual1 = (long)executor.OutputQueue.Dequeue();
+      var actual1 = executor.OutputQueue.Dequeue();
       executor.Initialize();
-      executor.InputQueue.Enqueue((long)0);
+      executor.InputQueue.Enqueue(0);
       executor.Execute();
-      var actual2 = (long)executor.OutputQueue.Dequeue();
+      var actual2 = executor.OutputQueue.Dequeue();
       executor.Initialize();
-      executor.InputQueue.Enqueue((long)8);
+      executor.InputQueue.Enqueue(8);
       executor.Execute();
-      var actual3 = (long)executor.OutputQueue.Dequeue();
+      var actual3 = executor.OutputQueue.Dequeue();
 
       Assert.AreEqual<long>(expected1, actual1, "OK!");
       Assert.AreEqual<long>(expected2, actual2, "OK!");
@@ -173,17 +173,17 @@ namespace IntCodeTest
       long expected2 = 1000;
       long expected3 = 1001;
 
-      executor.InputQueue.Enqueue((long)-3);
+      executor.InputQueue.Enqueue(-3);
       executor.Execute();
-      var actual1 = (long)executor.OutputQueue.Dequeue();
+      var actual1 = executor.OutputQueue.Dequeue();
       executor.Initialize();
-      executor.InputQueue.Enqueue((long)8);
+      executor.InputQueue.Enqueue(8);
       executor.Execute();
-      var actual2 = (long)executor.OutputQueue.Dequeue();
+      var actual2 = executor.OutputQueue.Dequeue();
       executor.Initialize();
-      executor.InputQueue.Enqueue((long)15);
+      executor.InputQueue.Enqueue(15);
       executor.Execute();
-      var actual3 = (long)executor.OutputQueue.Dequeue();
+      var actual3 = executor.OutputQueue.Dequeue();
 
       Assert.AreEqual<long>(expected1, actual1, "OK!");
       Assert.AreEqual<long>(expected2, actual2, "OK!");
@@ -208,7 +208,7 @@ namespace IntCodeTest
       long expected = 1219070632396864;
       var executor = new Executor(intCode);
       executor.Execute();
-      var actual = (long)executor.OutputQueue.Dequeue();
+      var actual = executor.OutputQueue.Dequeue();
       Assert.AreEqual<long>(expected, actual, "OK!");
     }
 
@@ -219,7 +219,7 @@ namespace IntCodeTest
       long expected = 1125899906842624;   
       var executor = new Executor(intCode);
       executor.Execute();
-      var actual = (long)executor.OutputQueue.Dequeue();
+      var actual = executor.OutputQueue.Dequeue();
       Assert.AreEqual<long>(expected, actual, "OK!");
     }
   }
